Issue #55: Standardize usages of long

+++ include/openenclave/bits/elf.h
@@ -4,6 +4,7 @@
 #ifndef _OE_ELF_H
 #define _OE_ELF_H
 
+#include <openenclave/types.h>
 #include <stddef.h>
 #include <stdio.h>
 #include <stdlib.h>
@@ -141,13 +142,13 @@ ELF_EXTERNC_BEGIN
 #define ELF64_R_TYPE(i) ((i)&0xffffffffL)
 #define ELF64_R_INFO(s, t) (((s) << 32) + ((t)&0xffffffffL))
 
-typedef unsigned long long Elf64_Addr;
-typedef unsigned long long Elf64_Off;
+typedef uint64_t Elf64_Addr;
+typedef uint64_t Elf64_Off;
 typedef unsigned short Elf64_Half;
 typedef unsigned int Elf64_Word;
 typedef signed int Elf64_Sword;
-typedef unsigned long long Elf64_Xword;
-typedef signed long long Elf64_Sxword;
+typedef uint64_t Elf64_Xword;
+typedef int64_t Elf64_Sxword;
 typedef signed int Elf64_Sword;
 
 typedef struct
